# openapi.yml
openapi: 3.1.0
info:
  title: DatabasesEverywhere API
  version: 0.10.0
  description: API used by panels/addons to create, manage, monitor, back up, and restore hosted database instances.
servers:
  - url: http://127.0.0.1:8090
    description: Local daemon API
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: panel token
  schemas:
...
    HeartbeatResponse:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status: {type: string, const: ok}
...
    CreateInstanceAcceptedResponse:
      type: object
      additionalProperties: false
      required: [instance_id, status, status_url]
      properties:
        instance_id: {type: string}
        status:
          type: string
          const: creating
        status_url:
          type: string
          description: Origin-relative URL for polling creation status and progress.
...
    ConfigPatchResponse:
      type: object
      required: [applied, restart_required, config_path]
      properties:
        applied:
          type: boolean
        restart_required:
          type: boolean
        config_path:
          type: string
...
  responses:
    ApiError:
      description: Request rejection or server-side failure. Internal details are logged under error_id and are never returned to clients.
      headers:
        X-Error-Id:
          description: Opaque support identifier included for server-side failures.
          schema: {type: string, format: uuid}
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing, invalid, or insufficient bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
paths:
  /api/system:
    get:
      summary: Get daemon runtime/configuration summary.
      x-required-scope: system:read
      responses:
        default:
          $ref: "#/components/responses/ApiError"
        "200":
          description: System status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/system/config:
    patch:
      summary: Merge a validated config patch into the daemon config file.
      description: Writes config.yml and returns restart_required=true. Refuses uuid, token_id, and token edits. The allocation.prevent_cpu_overallocation, allocation.prevent_memory_overallocation, and allocation.prevent_disk_overallocation booleans may be changed independently; all default to true.
      x-required-scope: config:admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        default:
          $ref: "#/components/responses/ApiError"
        "200":
          description: Config patch applied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigPatchResponse"
  /api/heartbeat:
    get:
      summary: Lightweight authenticated API liveness check.
      description: Returns success when the authenticated management API is reachable, independently of database instance and gateway readiness.
      x-required-scope: system:read
      responses:
        default:
          $ref: "#/components/responses/ApiError"
        "200":
          description: Heartbeat response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HeartbeatResponse"
...
  /api/instances:
    get:
      summary: List database instances with their live classified container statuses.
      x-required-scope: instances:read
      responses:
        default:
          $ref: "#/components/responses/ApiError"
        "200":
          description: Instance list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Instance"
    post:
      summary: Create a database instance.
      x-required-scope: instances:write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInstanceRequest"
      responses:
        default:
          $ref: "#/components/responses/ApiError"
        "202":
          description: Instance creation accepted and continuing in the background.
          headers:
            Location:
              description: Origin-relative instance creation status URL.
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateInstanceAcceptedResponse"
        "409":
          description: Instance ID already exists.
        "503":
          description: Creation admission is closed or the bounded creation queue is full.
  /api/instances/{instance_id}:
    parameters:
      - name: instance_id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get one instance with its live classified container status.
      x-required-scope: instances:read
      responses:
        default:
          $ref: "#/components/responses/ApiError"
        "200":
          description: Instance metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instance"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Irreversibly delete one instance, its managed container, jobs, artifacts, backups, recovery volumes, and all managed data.
      x-required-scope: instances:write
      parameters:
        - name: confirm
          in: query
          required: true
          description: Must be true because instance deletion always purges managed data.
          schema:
            type: boolean
            const: true
        - name: reason
          in: query
          required: true
          description: Audit reason for irreversible instance deletion; at most 512 characters.
          schema:
            type: string
            minLength: 1
            maxLength: 512
      responses:
        default:
          $ref: "#/components/responses/ApiError"
        "200":
          description: Delete result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteInstanceResponse"
  /api/instances/{instance_id}/status:
    get:
      summary: Get instance status.
      x-required-scope: instances:read
      parameters:
        - name: instance_id
          in: path
          required: true
          schema:
            type: string
      responses:
        default:
          $ref: "#/components/responses/ApiError"
        "200":
          description: Status response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstanceStatusResponse"
  /api/instances/{instance_id}/limits:
    patch:
      summary: Update CPU, memory, and disk limits for an instance.
      x-required-scope: instances:write
      parameters:
        - name: instance_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LimitsRequest"
      responses:
        default:
          $ref: "#/components/responses/ApiError"
        "200":
          description: Updated instance metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instance"
        "503":
          description: An enabled CPU, memory, or disk allocation guard rejected a limit increase. Limit decreases remain allowed.
  /api/instances/{instance_id}/password:
    patch:
      summary: Reset an instance tenant password or API key.
      description: Atomically rotates the database credential, recreates the managed container with its current image and existing data, validates the replacement credential, and attempts to restore the previous credential if any step fails. The instance must be running and experiences a brief connection interruption.
      x-required-scope: instances:write
      parameters:
        - name: instance_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetInstancePasswordRequest"
      responses:
        default:
          $ref: "#/components/responses/ApiError"
        "200":
          description: Credential reset result. The password is never included.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResetInstancePasswordResponse"
  /api/instances/{instance_id}/reconcile:
    post:
      summary: Reconcile stored instance status with the container runtime.
      x-required-scope: instances:write
      parameters:
        - name: instance_id
          in: path
          required: true
          schema:
            type: string
      responses:
        default:
          $ref: "#/components/responses/ApiError"
        "200":
          description: Reconciled instance status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstanceStatusResponse"
  /api/instances/{instance_id}/power:
    post:
      summary: Unified power API for start, stop, restart, and kill.
      x-required-scope: instances:write
      parameters:
        - name: instance_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PowerRequest"
      responses:
        default:
          $ref: "#/components/responses/ApiError"
        "200":
          description: Power action result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PowerResponse"
  /api/instances/{instance_id}/image:
    patch:
      summary: Pull and switch an instance to a new database image.
      description: Same-major updates recreate the container on the existing volume. Major version changes require major_upgrade=true and run the safer export/import migration workflow where supported.
      x-required-scope: instances:write
      parameters:
        - name: instance_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInstanceImageRequest"
      responses:
        default:
          $ref: "#/components/responses/ApiError"
        "200":
          description: Image update result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateInstanceImageResponse"
  /api/admin/resources:
    get:
      summary: List resource usage for all instances.
      x-required-scope: resources:admin
      responses:
        default:
          $ref: "#/components/responses/ApiError"
        "200":
          description: Resource reports.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResourceReport"
  /api/admin/resources/summary:
    get:
      summary: Get node capacity, allocation, and whole-host resource pressure.
      description: Allocation includes all managed instances, including stopped instances that may be restarted. Host usage covers the entire Linux node. Managed usage is null when any running or booting instance could not be sampled. Disk totals describe the filesystem backing paths.volumes.
      x-required-scope: resources:admin
      responses:
        default:
          $ref: "#/components/responses/ApiError"
        "200":
          description: Node-wide resource summary for scheduling and capacity decisions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeResourceSummary"
  /api/instances/{instance_id}/resources:
    get:
      summary: Get resource usage for one instance.
      x-required-scope: resources:read
      parameters:
        - name: instance_id
          in: path
          required: true
          schema:
            type: string
      responses:
        default:
          $ref: "#/components/responses/ApiError"
        "200":
          description: Resource report.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceReport"
  /api/instances/{instance_id}/logs:
    get:
      summary: Get recent logs for one instance.
      x-required-scope: logs:read
      parameters:
        - name: instance_id
          in: path
          required: true
          schema:
            type: string
        - name: tail
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 2000
            default: 200
      responses:
        default:
          $ref: "#/components/responses/ApiError"
        "200":
          description: Log output.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogsResponse"
  /api/instances/{instance_id}/export:
    post:
      summary: Export one instance to a protocol-native dump artifact.
      x-required-scope: import-export:write
      parameters:
        - name: instance_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportRequest"
      responses:
        default:
          $ref: "#/components/responses/ApiError"
        "202":
          description: Queued export job.
          headers:
            Location:
              description: Origin-relative job status URL.
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportExportJob"
  /api/instances/{instance_id}/import:
    post:
      summary: Import a protocol-native artifact or a remote database into one instance.
      x-required-scope: import-export:write
      parameters:
        - name: instance_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportRequest"
      responses:
        default:
          $ref: "#/components/responses/ApiError"
        "202":
          description: Queued import job.
          headers:
            Location:
              description: Origin-relative job status URL.
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportExportJob"
  /api/instances/{instance_id}/import-export/jobs:
    get:
      summary: List import/export jobs for one instance.
      x-required-scope: import-export:read
      parameters:
        - name: instance_id
          in: path
          required: true
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        default:
          $ref: "#/components/responses/ApiError"
        "200":
          description: Job list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ImportExportJob"
...
  /api/admin/backups/status:
    get:
      summary: Get automatic backup configuration status.
      x-required-scope: backups:admin
      responses:
        default:
          $ref: "#/components/responses/ApiError"
        "200":
          description: Backup status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackupStatusResponse"
...